StatsViewのコンポーネントを個別ファイルに分離
- GraphTypeSelectorを独立したファイルに移動
- ProficiencyRatioViewを独立したファイルに移動
- WeeklyTrendViewを独立したファイルに移動
- StatsViewを更新して分離したコンポーネントを使用するように修正
